Fix statpopgen GnomADBuilder to emit UInt8 GT to match schema
The "Use Uint8 for GT field" change set the GT field schema to
list(UInt8), but the Arrow builder still produced list(UInt64),
causing data generation to fail at write time:

  column types must match schema types, expected List(UInt8)
  but found List(UInt64)

Switch GT_builder to ListBuilder<UInt8Builder> and have
parse_genotype return Option<u8> (genotype dosage is only ever
NULL/0/1/2), so the produced arrays match the declared schema.

+++ vortex-bench/src/statpopgen/vcf_conversion.rs
@@ -128,7 +128,7 @@ pub fn value_list_string<'a>(
     })
 }
 
-pub fn parse_genotype(gt: Option<EntryValue>) -> VortexResult<Option<u64>> {
+pub fn parse_genotype(gt: Option<EntryValue>) -> VortexResult<Option<u8>> {
     let Some(gt) = gt else {
         return Ok(None);
     };
@@ -140,7 +140,7 @@ pub fn parse_genotype(gt: Option<EntryValue>) -> VortexResult<Option<u64>> {
         .process_results(|iter| iter.map(|x| x.0).collect::<Vec<_>>())?[..]
     {
         [None, None] => Ok(None),
-        [Some(l), Some(r)] => Ok(Some(l as u64 + r as u64)),
+        [Some(l), Some(r)] => Ok(Some(l as u8 + r as u8)),
         _ => vortex_bail!("wtf {:?}", gt),
     }
 }
